# api-spec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "CS2KZ API",
    "description": "Source Code available on [GitHub](https://github.com/KZGlobalTeam/cs2kz-api).",
    "contact": {
      "name": "AlphaKeks",
      "email": "alphakeks@dawn.sh"
    },
    "license": {
      "name": "Licensed under the GPLv3",
      "url": "https://www.gnu.org/licenses/gpl-3.0"
    },
    "version": "0.0.0"
  },
  "paths": {
...
    "/players/{player}": {
      "get": {
        "tags": [
          "Players"
        ],
        "operationId": "get",
        "parameters": [
          {
            "name": "player",
            "in": "path",
            "required": false,
            "schema": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/SteamID"
                },
                {
                  "type": "string",
                  "description": "A player's name."
                }
              ],
              "description": "Players are usually identified by their [SteamID], which is unique, or their name."
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "A KZ player.",
                  "required": [
                    "name",
                    "steam_id",
                    "is_banned"
                  ],
                  "properties": {
                    "name": {
                      "type": "string",
                      "description": "The player's name."
                    },
                    "steam_id": {
                      "$ref": "#/components/schemas/SteamID"
                    },
                    "ip_address": {
                      "type": "string",
                      "description": "The player's IP address.",
                      "nullable": true
                    },
                    "is_banned": {
                      "type": "boolean",
                      "description": "Whether the player is currently banned."
                    }
                  }
                }
              }
            }
          },
          "204": {
            "description": ""
          },
          "400": {
            "description": ""
          },
          "500": {
            "description": ""
          }
        }
      }
    },
    "/players/{steam_id}": {
      "patch": {
        "tags": [
          "Players"
        ],
        "operationId": "patch",
        "parameters": [
          {
            "name": "steam_id",
            "in": "path",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "title": "Steam ID",
                  "example": "STEAM_1:1:161178172"
                },
                {
                  "type": "string",
                  "title": "Steam ID3",
                  "example": "U:1:322356345"
                },
                {
                  "type": "integer",
                  "title": "Steam ID32",
                  "example": 322356345
                },
                {
                  "type": "integer",
                  "title": "Steam ID64",
                  "example": 76561198282622073
                }
              ],
              "description": "a player's SteamID",
              "example": "STEAM_1:1:161178172"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlayerUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": ""
          },
          "400": {
            "description": ""
          },
          "401": {
            "description": ""
          },
          "422": {
            "description": ""
          },
          "500": {
            "description": ""
          }
        },
        "security": [
          {
            "CS2 Server": []
          }
        ]
      }
    },
...
    "/maps": {
      "get": {
        "tags": [
          "Maps"
        ],
        "operationId": "get",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "A KZ map.\n\nThe only reason this is named `FullMap` instead of just `Map`, is because `utoipa` macros are\nstupid.",
                  "required": [
                    "id",
                    "name",
                    "global_status",
                    "workshop_id",
                    "checksum",
                    "mappers",
                    "courses",
                    "created_on"
                  ],
                  "properties": {
                    "id": {
                      "type": "integer",
                      "format": "uint16",
                      "description": "The map's ID.",
                      "minimum": 0
                    },
                    "name": {
                      "type": "string",
                      "description": "The map's name."
                    },
                    "description": {
                      "type": "string",
                      "description": "The map's description.",
                      "nullable": true
                    },
                    "global_status": {
                      "$ref": "#/components/schemas/GlobalStatus"
                    },
                    "workshop_id": {
                      "type": "integer",
                      "format": "uint32",
                      "description": "The map's workshop ID.",
                      "minimum": 0
                    },
                    "checksum": {
                      "type": "integer",
                      "format": "uint32",
                      "description": "The map's checksum.",
                      "minimum": 0
                    },
                    "mappers": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Player"
                      },
                      "description": "List of players who have contributed to the creation of this map."
                    },
                    "courses": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Course"
                      },
                      "description": "List of courses on this map."
                    },
                    "created_on": {
                      "type": "string",
                      "format": "date-time",
                      "description": "When this map was approved."
                    }
                  }
                }
              }
            }
          },
          "204": {
            "description": ""
          },
          "400": {
            "description": ""
          },
          "500": {
            "description": ""
          }
        }
      },
      "put": {
        "tags": [
          "Maps"
        ],
        "operationId": "put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewMap"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "A newly created map.",
                  "required": [
                    "map_id"
                  ],
                  "properties": {
                    "map_id": {
                      "type": "integer",
                      "format": "uint16",
                      "description": "The map's ID.",
                      "minimum": 0
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": ""
          },
          "401": {
            "description": ""
          },
          "422": {
            "description": ""
          },
          "500": {
            "description": ""
          }
        },
        "security": [
          {
            "Browser Session": [
              "maps"
            ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Course": {
        "type": "object",
        "description": "A map course.",
        "required": [
          "id",
          "mappers",
          "filters"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "uint32",
            "description": "The course's ID.",
            "minimum": 0
          },
          "name": {
            "type": "string",
            "description": "The course's name.",
            "nullable": true
          },
          "description": {
            "type": "string",
            "description": "The course's description.",
            "nullable": true
          },
          "mappers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Player"
            },
            "description": "List of players who have contributed to the creation of this course."
          },
          "filters": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Filter"
            },
            "description": "The course's filters."
          }
        }
      },
...
      "Filter": {
        "type": "object",
        "description": "A course filter.",
        "required": [
          "id",
          "mode",
          "teleports",
          "tier",
          "ranked_status"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "uint32",
            "description": "The filter's ID.",
            "minimum": 0
          },
          "mode": {
            "$ref": "#/components/schemas/Mode"
          },
          "teleports": {
            "type": "boolean",
            "description": "The \"runtype\" this filter applies to (whether teleports are used or not)."
          },
          "tier": {
            "$ref": "#/components/schemas/Tier"
          },
          "ranked_status": {
            "$ref": "#/components/schemas/RankedStatus"
          },
          "notes": {
            "type": "string",
            "description": "Extra notes about this filter.",
            "nullable": true
          }
        }
      },
...
      "GlobalStatus": {
        "anyOf": [
          {
            "type": "string",
            "title": "Name",
            "enum": [
              "not_global",
              "in_testing",
              "global"
            ],
            "example": "global"
          },
          {
            "type": "integer",
            "title": "Integer",
            "enum": [
              -1,
              0,
              1
            ],
            "example": 1
          }
        ],
        "default": "global",
        "example": "global"
      },
...
      "Mode": {
        "anyOf": [
          {
            "type": "string",
            "title": "Name",
            "enum": [
              "vanilla",
              "classic"
            ],
            "example": "classic"
          },
          {
            "type": "integer",
            "title": "ID",
            "enum": [
              1,
              2
            ],
            "example": 1
          }
        ],
        "example": "classic"
      },
...
      "NewCourse": {
        "type": "object",
        "description": "Request body for submitting new map courses.",
        "required": [
          "mappers",
          "filters"
        ],
        "properties": {
          "name": {
            "type": "string",
            "description": "The course's name.",
            "nullable": true
          },
          "description": {
            "type": "string",
            "description": "The course's description.",
            "nullable": true
          },
          "mappers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SteamID"
            },
            "description": "List of players who have contributed to the creation of this course."
          },
          "filters": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NewFilter"
            },
            "description": "The course's filters."
          }
        }
      },
      "NewFilter": {
        "type": "object",
        "description": "Request body for submitting new course filters.",
        "required": [
          "mode",
          "teleports",
          "tier",
          "ranked_status"
        ],
        "properties": {
          "mode": {
            "$ref": "#/components/schemas/Mode"
          },
          "teleports": {
            "type": "boolean",
            "description": "The \"runtype\" this filter applies to (whether teleports are used or not)."
          },
          "tier": {
            "$ref": "#/components/schemas/Tier"
          },
          "ranked_status": {
            "$ref": "#/components/schemas/RankedStatus"
          },
          "notes": {
            "type": "string",
            "description": "Extra notes about this filter.",
            "nullable": true
          }
        }
      },
...
      "NewMap": {
        "type": "object",
        "description": "Request body for submitting new maps.",
        "required": [
          "workshop_id",
          "global_status",
          "mappers",
          "courses"
        ],
        "properties": {
          "workshop_id": {
            "type": "integer",
            "format": "uint32",
            "description": "The map's workshop ID.",
            "minimum": 0
          },
          "description": {
            "type": "string",
            "description": "The map's description.",
            "nullable": true
          },
          "global_status": {
            "$ref": "#/components/schemas/GlobalStatus"
          },
          "mappers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SteamID"
            },
            "description": "List of players who have contributed to the creation of this map."
          },
          "courses": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NewCourse"
            },
            "description": "List of courses on this map."
          }
        }
      },
...
      "Player": {
        "type": "object",
        "description": "A KZ player.",
        "required": [
          "name",
          "steam_id"
        ],
        "properties": {
          "name": {
            "type": "string",
            "description": "The player's name."
          },
          "steam_id": {
            "$ref": "#/components/schemas/SteamID"
          }
        }
      },
...
      "PlayerUpdate": {
        "type": "object",
        "description": "Request body for updating players.",
        "required": [
          "name",
          "ip_address",
          "session"
        ],
        "properties": {
          "name": {
            "type": "string",
            "description": "The player's name."
          },
          "ip_address": {
            "type": "string",
            "description": "The player's IP address."
          },
          "session": {
            "$ref": "#/components/schemas/Session"
          }
        }
      },
...
      "RankedStatus": {
        "anyOf": [
          {
            "type": "string",
            "title": "Name",
            "enum": [
              "never",
              "unranked",
              "ranked"
            ],
            "example": "ranked"
          },
          {
            "type": "integer",
            "title": "Integer",
            "enum": [
              -1,
              0,
              1
            ],
            "example": 1
          }
        ],
        "example": "ranked"
      },
...
      "SteamID": {
        "anyOf": [
          {
            "type": "string",
            "title": "Steam ID",
            "example": "STEAM_1:1:161178172"
          },
          {
            "type": "string",
            "title": "Steam ID3",
            "example": "U:1:322356345"
          },
          {
            "type": "integer",
            "title": "Steam ID32",
            "example": 322356345
          },
          {
            "type": "integer",
            "title": "Steam ID64",
            "example": 76561198282622073
          }
        ],
        "description": "a player's SteamID",
        "example": "STEAM_1:1:161178172"
      },
...
      "Tier": {
        "anyOf": [
          {
            "type": "string",
            "title": "Name",
            "enum": [
              "very_easy",
              "easy",
              "medium",
              "advanced",
              "hard",
              "very_hard",
              "extreme",
              "death",
              "unfeasible",
              "impossible"
            ],
            "example": "hard"
          },
          {
            "type": "integer",
            "title": "Numeric Value",
            "enum": [
              1,
              2,
              3,
              4,
              5,
              6,
              7,
              8,
              9,
              10
            ],
            "example": 1
          }
        ],
        "example": "hard"
      },
...
    },
    "securitySchemes": {
      "API Key": {
        "type": "http",
        "scheme": "bearer"
      },
      "Browser Session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "kz-auth"
      },
      "CS2 Server": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}